pruebas + pruebas fix + fix en generar (logo path) + loqueras de composer

+++ modelo/calendario.php
@@ -86,25 +86,21 @@ class calendario extends Conexion
     PRIVATE function modificar_d() {
         try {
             // Validaciones para los datos del evento
-            $this->validar_conexion($this->con);
             Validaciones::alfanumerico($this->descripcion, "1,200", "La descripción no es válida");
             Validaciones::fecha($this->fecha, "Fecha del evento no válida");
             Validaciones::validar($this->recurrente,"/^(?:0|1)$/","El valor de recurrencia no es valido");
+            $this->validar_conexion($this->con);
+            $this->con->beginTransaction();
     
             // Verificar si ya existe un evento con la misma fecha
-            $consultaDuplicado = $this->con->prepare("SELECT 1 FROM `calendario` WHERE `fecha` = :fecha AND `fecha` != :fecha_actual");
-            $consultaDuplicado->bindValue(":fecha", $this->fecha);
-            $consultaDuplicado->bindValue(":fecha_actual", $this->fecha); // Actual fecha del evento
-            $consultaDuplicado->execute();
-            
-            if ($consultaDuplicado->fetchColumn() > 0) {
-                
-                $r['resultado'] = 'error';
-                $r['mensaje'] = "Ya existe un evento en la fecha especificada";
-                return $r;
+            $consulta = $this->con->prepare("SELECT 1 FROM `calendario` WHERE `fecha` = :fecha");
+            $consulta->bindValue(":fecha", $this->fecha);
+            $consulta->execute();
+
+            if(!$consulta->fetch()){
+                throw new Exception("La fecha a elimnar no existe",1);
             }
     
-            $this->con->beginTransaction();
     
             $consulta = $this->con->prepare("UPDATE calendario SET descripcion = :descripcion, recurrente = :recurrente WHERE fecha = :fecha");
             $consulta->bindValue(":descripcion", $this->descripcion);
@@ -138,6 +134,14 @@ class calendario extends Conexion
     
             $this->validar_conexion($this->con);
             $this->con->beginTransaction();
+
+            $consulta = $this->con->prepare("SELECT 1 FROM `calendario` WHERE `fecha` = :fecha");
+            $consulta->bindValue(":fecha", $this->fecha);
+            $consulta->execute();
+
+            if(!$consulta->fetch()){
+                throw new Exception("La fecha a elimnar no existe",1);
+            }
     
             $consulta = $this->con->prepare("DELETE FROM `calendario` WHERE `fecha` = :fecha");
             $consulta->bindValue(":fecha", $this->fecha);
@@ -167,6 +171,11 @@ class calendario extends Conexion
             // Validaciones para obtener eventos
             Validaciones::numero($year, "4", "El año no es válido");
             Validaciones::numero($month, "1,2", "El mes no es válido");
+
+            // verifica que el mes no sea mayor a 12 ni menor a 1    
+            if (intval($month) < 1 || intval($month) > 12) {
+                throw new Exception("El mes no es valido",1);
+            }
     
             $this->validar_conexion($this->con);
             $this->con->beginTransaction();
